Port process check to pyest
