+++ process/requirements-dev.txt
@@ -1,0 +1,2 @@
+pytest
+mock
