+++ process/tox.ini
@@ -1,0 +1,22 @@
+[tox]
+minversion = 2.0
+basepython = py27
+envlist =
+    process
+    flake8
+
+[testenv]
+platform = linux2|darwin|windows
+
+[testenv:process]
+deps = 
+    ../datadog_checks_base
+    -rrequirements-dev.txt
+commands =
+    pip install --require-hashes -r requirements.txt
+    pytest .
+
+[testenv:flake8]
+skip_install = true
+deps = flake8
+commands = flake8 .
